# api/app_management/openapi.yaml
openapi: 3.0.3
info:
  title: NoxOS App Management API
  version: v2
  description: |
    <picture>
        <source media="(prefers-color-scheme: dark)" srcset="https://raw.githubusercontent.com/Nox-OS/logo/main/noxos/noxos_banner_dark_night_800px.png">
        <source media="(prefers-color-scheme: light)" srcset="https://raw.githubusercontent.com/Nox-OS/logo/main/noxos/noxos_banner_twilight_blue_800px.png">
        <img alt="NoxOS" src="https://raw.githubusercontent.com/Nox-OS/logo/main/noxos/noxos_banner_twilight_blue_800px.png">
    </picture>

    NoxOS App Management manages app lifecycle, including installing, uninstalling, starting, stopping, setting, etc.

    For issues and discussions, please visit the [GitHub repository](https://github.com/Nox-OS/NoxOS) or join [our Discord](https://discord.gg/knqAbbBbeX).
servers:
  - url: /v2/app_management
...
security:
  - access_token: []
paths:
...
  /convert:
    post:
      summary: Convert to compose app from another file format
      operationId: convert
      tags:
        - Common methods
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum:
              - appfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          $ref: "#/components/responses/ComposeAppOK"
        "400":
          $ref: "#/components/responses/ResponseBadRequest"
...
  /appstore:
    get:
      summary: Get the list of registered app stores
      description: |
        (TODO)
      operationId: appStoreList
      tags:
        - AppStore methods
      responses:
        "200":
          $ref: "#/components/responses/AppStoreListOK"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"

    post:
      summary: Register an app store
      description: |
        (TODO)
      operationId: registerAppStore
      tags:
        - AppStore methods
      parameters:
        - $ref: "#/components/parameters/AppStoreURL"
      responses:
        "200":
          $ref: "#/components/responses/AppStoreRegisterOK"
        "400":
          $ref: "#/components/responses/ResponseBadRequest"
        "409":
          $ref: "#/components/responses/ResponseConflict"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"
  /appstore/{id}:
    delete:
      summary: Unregister an app store
      description: |
        (TODO)
      operationId: unregisterAppStore
      tags:
        - AppStore methods
      parameters:
        - $ref: "#/components/parameters/AppStoreID"
      responses:
        "200":
          $ref: "#/components/responses/AppStoreUnregisterOK"
        "404":
          $ref: "#/components/responses/ResponseNotFound"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"
...
  /apps:
    get:
      summary: Get app list from registered app stores
      description: |
        Get lists of StoreInfo of all compose apps from AppStore.
      operationId: composeAppStoreInfoList
      tags:
        - AppStore methods
      parameters:
        - $ref: "#/components/parameters/StoreAppCategoryFilter"
        - $ref: "#/components/parameters/StoreAppAuthorTypeFilter"
        - $ref: "#/components/parameters/StoreAppRecommendedFilter"
      responses:
        "200":
          $ref: "#/components/responses/ComposeAppStoreInfoListsOK"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"
  /apps/upgradable:
    get:
      summary: Get the list of upgradable apps from installed app.
      description: |
        (TODO)
      operationId: upgradableAppList
      tags:
        - AppStore methods
      responses:
        "200":
          $ref: "#/components/responses/UpgradableAppListOK"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"
  /apps/{id}:
    get:
      summary: Get app info from registered app stores
      description: |
        Get StoreInfo of specific compose app.

        > A compose app contains one or more container `apps`. One of the `apps` is the `main` of this compose app, whereas others are dependency `apps`.
      operationId: composeAppStoreInfo
      tags:
        - AppStore methods
      parameters:
        - $ref: "#/components/parameters/StoreAppIDString"
      responses:
        "200":
          $ref: "#/components/responses/ComposeAppStoreInfoOK"
        "404":
          $ref: "#/components/responses/ResponseNotFound"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"
  /apps/{id}/stable:
    get:
      summary: Get app stable version from registered app stores
      description: |
        Get stable version(Official Store Tag) of main service of specific compose app.
      operationId: composeAppMainStableTag
      tags:
        - AppStore methods
      parameters:
        - $ref: "#/components/parameters/StoreAppIDString"
      responses:
        "200":
          $ref: "#/components/responses/ComposeAppStoreTagOK"
        "404":
          $ref: "#/components/responses/ResponseNotFound"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"
  /apps/{id}/stable/{serviceName}:
    get:
      summary: Get server of app stable version from registered app stores
      description: |
        Get stable version(Official Store Tag) of main service of specific compose app.
      operationId: composeAppServiceStableTag
      tags:
        - AppStore methods
      parameters:
        - $ref: "#/components/parameters/StoreAppIDString"
        - name: serviceName
          in: path
          description: Service name of the compose app
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/ComposeAppStoreTagOK"
        "404":
          $ref: "#/components/responses/ResponseNotFound"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"
  /apps/{id}/compose:
    get:
      summary: Get compose details of an app from registered app stores
      description: |
        Accept both `application/json` and `application/yaml` content type for the `Accept` header in the HTTP request.

        If the `Accept` header is set to `application/yaml`, the response will be the original `docker-compose.yml` content that follows [Compose Specification](https://compose-spec.io).
      operationId: composeApp
      tags:
        - AppStore methods
      parameters:
        - $ref: "#/components/parameters/StoreAppIDString"
      responses:
        "200":
          $ref: "#/components/responses/ComposeAppOK"
        "404":
          $ref: "#/components/responses/ResponseNotFound"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"
...
  /web/sync/appstore:
    post:
      summary: Register an app store in Sync.
      description: |
        (TODO)
      operationId: registerAppStoreSync
      tags:
        - AppStore methods
      parameters:
        - $ref: "#/components/parameters/AppStoreURL"
      responses:
        "200":
          $ref: "#/components/responses/AppStoreRegisterOK"
        "400":
          $ref: "#/components/responses/ResponseBadRequest"
        "409":
          $ref: "#/components/responses/ResponseConflict"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"
components:
  securitySchemes:
    access_token:
      type: apiKey
      in: header
      name: Authorization
  parameters:
...
    AppStoreURL:
      name: url
      description: URL of the app store
      in: query
      schema:
        type: string
        example: "https://github.com/Nox-OS/NoxOS-AppStore"
    AppStoreID:
      name: id
      description: ID of the app store
      in: path
      required: true
      schema:
        type: integer
        example: 0
    StoreAppIDString:
      name: id
      description: Store app ID of a compose app (unique across all app stores)
      in: path
      required: true
      schema:
        type: string
        example: "syncthing"
    StoreAppCategoryFilter:
      name: category
      description: Category of the store app
      in: query
      schema:
        type: string
        example: "chat"
    StoreAppAuthorTypeFilter:
      name: author_type
      description: Author type of the store app
      in: query
      schema:
        $ref: "#/components/schemas/StoreAppAuthorType"
    StoreAppRecommendedFilter:
      name: recommend
      description: The store app is recommended
      in: query
      schema:
        type: boolean
        default: false
...
  responses:
...
    ResponseInternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseResponse"
          example:
            message: "Internal Server Error"
    ResponseNotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseResponse"
          example:
            message: "Not Found"
    ResponseBadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseResponse"
          example:
            message: "Bad Request"
    ResponseConflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseResponse"
          example:
            message: "Conflict"
...
    UpgradableAppListOK:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/BaseResponse"
              - properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/UpgradableAppInfo"
    AppStoreListOK:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/BaseResponse"
              - properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AppStoreMetadata"
    AppStoreRegisterOK:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/BaseResponse"
    AppStoreUnregisterOK:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/BaseResponse"
...
    ComposeAppStoreInfoListsOK:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/BaseResponse"
              - properties:
                  data:
                    $ref: "#/components/schemas/ComposeAppStoreInfoLists"
    ComposeAppStoreTagOK:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/BaseResponse"
              - properties:
                  data:
                    $ref: "#/components/schemas/ComposeAppStoreTag"
    ComposeAppStoreInfoOK:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/BaseResponse"
              - properties:
                  data:
                    $ref: "#/components/schemas/ComposeAppStoreInfo"
...
    ComposeAppOK:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/BaseResponse"
              - properties:
                  data:
                    $ref: "#/components/schemas/ComposeAppWithStoreInfo"
        application/yaml:
          schema:
            type: string
...
  schemas:
    BaseResponse:
      properties:
        message:
          description: message returned by server side if there is any
          type: string
          example: ""
...
    AppStoreMetadata:
      properties:
        id:
          type: integer
          minimum: 0
          readOnly: true
          x-go-name: ID
        url:
          type: string
          readOnly: true
          example: "https://github.com/Nox-OS/NoxOS-AppStore/archive/refs/heads/main.zip"
          x-go-name: URL
        store_root:
          type: string
          readOnly: true
          example: "NoxOS-AppStore-main"
...
    ComposeAppStoreInfoLists:
      properties:
        installed:
          type: array
          items:
            $ref: "#/components/schemas/StoreAppID"
        list:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ComposeAppStoreInfo"
    ComposeAppStoreTag:
      required:
        - tag
      properties:
        tag:
          type: string
          example: "v10.2"
...
    ComposeApp:
      type: object
      description: |-
        See [Compose Specification](https://compose-spec.io) for the schema structure of `ComposeApp`.

      x-go-type: types.Project
      x-go-type-import:
        name: types
        path: github.com/compose-spec/compose-go/types
    ComposeAppWithStoreInfo:
      description: |-
        Schema for `ComposeAppOK` response.

        > Do not mistake it as `ComposeAppStoreInfo` which is just store information of a compose app.
      properties:
        store_info:
          $ref: "#/components/schemas/ComposeAppStoreInfo"
        compose:
          $ref: "#/components/schemas/ComposeApp"
        status:
          type: string
          readOnly: true
          example: running
        update_available:
          readOnly: true
          type: boolean
          example: true
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
        is_uncontrolled:
          readOnly: true
          title: is the app uncontrolled. 
          description: |
            the mean the image tag is custom. not stable or latest
          type: boolean
          example: false
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
...
    UpgradableAppInfo:
      title: Upgradable app info
      description: |-
        Schema for `UpgradableAppListOK` response.
      required:
        - title
        - version
        - store_app_id
        - status
        - icon
      properties:
        title:
          type: string
          example: "{en_US:\"Syncthing\"}"
        version:
          description: current version of the installed app
          type: string
          example: "v10.2"
        store_app_id:
          $ref: "#/components/schemas/StoreAppID"
        icon:
          type: string
          example: "https://cdn.jsdelivr.net/gh/Nox-OS/NoxOS-AppStore@main/Apps/Syncthing/icon.png"
        status:
          type: string
          enum:
            - "idle"
            - "updating"
...
    StoreAppID:
      title: Store app ID
      description: |
        Store app ID of the compose app (unique across all app stores)

        > The `store_app_id` might be same as app name most of the time. However app name might change, due to cloning for example.
        > If app name has changed, there is no way to associate it with the original app in the app store, unless we use the `store_app_id`.
      readOnly: true
      type: string
      example: "syncthing"
      x-go-name: StoreAppID
      x-oapi-codegen-extra-tags:
        mapstructure: "store_app_id"
        yaml: "store_app_id,omitempty"
    StoreAppAuthorType:
      title: Store app author type
      description: |
        Type of the author of the app

        - `official`: The app is contributed by the app owner, i.e. `author`` and `developer`` are the same.
        - `by_noxos`: The app is contributed by NoxOS team, i.e. `author` is `NoxOS Team`
        - `community`: The app is contributed by the community, i.e. any app that is not `official` or `by_noxos`
      readOnly: true
      type: string
      enum:
        - official
        - by_noxos
        - community
        - unknown
      x-enum-varnames:
        - AuthorTypeOfficial
        - AuthorTypeByNoxOS
        - AuthorTypeCommunity
        - AuthorTypeUnknown
...